Confirm deletes of log drains and certificates
Closes https://github.com/aptible/dashboard.aptible.com/issues/514

+++ app/certificates/route.js
@@ -17,6 +17,10 @@ export default Ember.Route.extend({
 
   actions: {
     delete: function(model){
+      // Confirm...
+      let confirmMsg = `\nAre you sure you want to delete ${model.get('commonName')}?\n`;
+      if (!confirm(confirmMsg)) { return false; }
+
       let stack = model.get('stack');
 
       model.deleteRecord();
@@ -29,4 +33,3 @@ export default Ember.Route.extend({
     }
   }
 });
-
